slint::docs

Module cargo_features

Feature flags and backend selection.

Use the following feature flags in your Cargo.toml to enable additional features.

compat-1-2 (enabled by default) — Mandatory feature: This feature is required to keep the compatibility with Slint 1.2 Newer patch version may put current functionality behind a new feature that would be enabled by default only if this feature was added. More info in this blog post

std (enabled by default) — Enable use of the Rust standard library.

gettext — Enable the translations using gettext

the @tr(...) code from .slint files will be transformed into call to dgettext with the crate name as domain name

translations must be enabled with the init_translations! macro

libm — This feature enables floating point arithmetic emulation using the libm crate. Use this in MCU environments where the processor does not support floating point arithmetic.

log — If enabled, calls of debug() in .slint files use to the log::debug!() macro of the log crate instead of just println!().

serde — Implement the serde::Serialize and serde::Deserialize for some of the base types such as SharedString and SharedVector.

unsafe-single-threaded — Slint uses internally some thread_local state.

When the std feature is enabled, Slint can use std::thread_local!, but when in a #![no_std] environment, we need a replacement. Using this feature, Slint will just use static variable disregarding Rust’s Send and Sync safety

Safety : You must ensure that there is only one single thread that call into the Slint API

accessibility (enabled by default) — Enable integration with operating system provided accessibility APIs

Enabling this feature will try to expose the tree of UI elements to OS provided accessibility APIs to support screen readers and other assistive technologies.

raw-window-handle-06 — Enable integration with raw-window-handle version 0.6. This provides a Window::window_handle() function that returns a struct that implements HasWindowHandle and HasDisplayHandle implementation.

image-default-formats — Enable the default image formats from the image crate, to support additional image formats in Image::load_from_path and @image-url. When this feature is disabled, only PNG and JPEG are supported. When enabled, the following image formats are supported: AVIF, BMP, DDS, Farbfeld, GIF, HDR, ICO, JPEG, EXR, PNG, PNM, QOI, TGA, TIFF, WebP.

live-preview — Enable the live preview feature

Enable this feature to reload the .slint files at runtime and reload it whenever the files are modified on disk. For this feature to work, it’s also required to set the SLINT_LIVE_PREVIEW environment variable during application compilation.

This is a feature for debugging and development. It’s not recommended to add this feature to your Cargo.toml. Instead, use the --features command line argument like so when building your app:

SLINT_LIVE_PREVIEW=1 cargo build --features slint/live-preview

Backends

Slint needs a backend that will act as liaison between Slint and the OS. By default, Slint will use the Qt backend, if Qt is installed, otherwise, it will use Winit with FemtoVG. Both backends are compiled in. If you want to not compile one of these you need to disable the default feature and re-enable one backend. It is also possible to use Slint without backend if you provide the platform abstraction yourself with platform::set_platform().

If you enable the Winit backend, you need to also include a renderer. renderer-femtovg is the default renderer.

It is also possible to select the backend and renderer at runtime when several are enabled, using the SLINT_BACKEND environment variable.

SLINT_BACKEND=Qt selects the Qt backend

SLINT_BACKEND=winit selects the winit backend

SLINT_BACKEND=winit-femtovg selects the winit backend with the FemtoVG renderer

SLINT_BACKEND=winit-skia selects the winit backend with the skia renderer

SLINT_BACKEND=winit-software selects the winit backend with the software renderer

If the selected backend is not available, the default will be used.

Here are the cargo features controlling the backend:

backend-qt — The Qt backend feature uses Qt for the windowing system integration and rendering. This backend also provides the native style. It requires Qt 5.15 or later to be installed. If Qt is not installed, the backend will not be operational

backend-winit — The winit crate is used for the event loop and windowing system integration. It supports Windows, macOS, web browsers, X11 and Wayland. X11 and wayland are only available when compiling for Linux or other Unix-like operating systems. With this feature, both X11 and Wayland are supported. For a smaller build, omit this feature and select one of the other specific backend-winit-XX features.

backend-winit-x11 — Simliar to backend-winit this enables the winit based event loop but only with support for the X Window System on Unix.

backend-winit-wayland — Simliar to backend-winit this enables the winit based event loop but only with support for the Wayland window system on Unix.

backend-default (enabled by default) — Alias to a backend and renderer that depends on the platform. Will select the Qt backend on linux if present, and the winit otherwise

renderer-femtovg (enabled by default) — Render using the FemtoVG crate.

renderer-femtovg-wgpu — Render using the FemtoVG crate and WGPU.

renderer-skia — Render using Skia.

renderer-skia-opengl — Same as renderer-skia, but Skia will always use OpenGL.

renderer-skia-vulkan — Same as renderer-skia, but Skia will always use Vulkan.

renderer-software (enabled by default) — Render using the software renderer.

backend-linuxkms — KMS with Vulkan or EGL and libinput on Linux are used to render the application in full screen mode, without any windowing system. Requires libseat. If you don’t have libseat, select backend-linuxkms-noseat instead.

backend-linuxkms-noseat — KMS with Vulkan or EGL and libinput on Linux are used to render the application in full screen mode, without any windowing system.

backend-android-activity-06 — Use the backend based on the android-activity crate. (Using it’s native activity feature)

backend-android-activity-05 — Deprecated Use previous version of android-activity. This is mutually exclusive with backend-android-activity-06.

unstable-wgpu-26 — Enable support for WGPU based rendering and expose WGPU based APIs based on WGPU version 26.x.

APIs guarded with this feature are NOT subject to the usual Slint API stability guarantees, because WGPU releases new major versions frequently. This feature as well as the APIs changed or removed in future minor releases of Slint, likely to be replaced by a feature with a similar name but the WGPU version suffix being bumped.

To avoid unintended compilation failures, we recommend to use the tilde requirement in your Cargo.toml when enabling this feature:

slint = { version = "~1.14", features = ["unstable-wgpu-26"] }

unstable-wgpu-27 — Enable support for WGPU based rendering and expose WGPU based APIs based on WGPU version 27.x.

APIs guarded with this feature are NOT subject to the usual Slint API stability guarantees, because WGPU releases new major versions frequently. This feature as well as the APIs changed or removed in future minor releases of Slint, likely to be replaced by a feature with a similar name but the WGPU version suffix being bumped.

To avoid unintended compilation failures, we recommend to use the tilde requirement in your Cargo.toml when enabling this feature:

slint = { version = "~1.14", features = ["unstable-wgpu-27"] }

unstable-winit-030 — APIs guarded with this feature are NOT subject to the usual Slint API stability guarantees, because winit releases new major versions more frequently than Slint. This feature as well as the APIs changed or removed in future minor releases of Slint, likely to be replaced by a feature with a similar name but the winit version suffix being bumped.

To avoid unintended compilation failures, we recommend to use the tilde requirement in your Cargo.toml when enabling this feature:

slint = { version = "~1.12", features = ["unstable-winit-030"] }

unstable-libinput-09 — Enable support for exposing libinput related APIs in the LinuxKMS backend.

APIs guarded with this feature are NOT subject to the usual Slint API stability guarantees. This feature as well as the APIs changed or removed in future minor releases of Slint, likely to be replaced by a feature with a similar name but the input version suffix being bumped.

To avoid unintended compilation failures, we recommend to use the tilde requirement in your Cargo.toml when enabling this feature:

slint = { version = "~1.13", features = ["unstable-libinput-09"] }

More information about the backend and renderers is available in the Slint Documentation“)]