Module slint::docs::generated_code

This module exists only to explain the API of the code generated from .slint design markup. Its described structure is not really contained in the compiled crate.

Structs

SampleComponent

This an example of the API that is generated for a component in .slint design markup. This may help you understand what functions you can call and how you can pass data in and out. This is the source code: