Function slint::run_event_loop

pub fn run_event_loop()

Enters the main event loop. This is necessary in order to receive events from the windowing system in order to render to the screen and react to user input.