Macro slint::include_modules

macro_rules! include_modules {
    () => { ... };
}

Include the code generated with the slint-build crate from the build script. After calling slint_build::compile in your build.rs build script, the use of this macro includes the generated Rust code and makes the exported types available for you to instantiate.

Check the documentation of the slint-build crate for more information.