slint::platform

Enum Key

#[non_exhaustive]
pub enum Key {

    Backspace,
    Tab,
    Return,
    Escape,
    Backtab,
    Delete,
    Shift,
    Control,
    Alt,
    AltGr,
    CapsLock,
    ShiftR,
    ControlR,
    Meta,
    MetaR,
    Space,
    UpArrow,
    DownArrow,
    LeftArrow,
    RightArrow,
    F1,
    F2,
    F3,
    F4,
    F5,
    F6,
    F7,
    F8,
    F9,
    F10,
    F11,
    F12,
    F13,
    F14,
    F15,
    F16,
    F17,
    F18,
    F19,
    F20,
    F21,
    F22,
    F23,
    F24,
    Insert,
    Home,
    End,
    PageUp,
    PageDown,
    ScrollLock,
    Pause,
    SysReq,
    Stop,
    Menu,
    Back,

}

The Key enum is used to map a specific key by name e.g. Key::Control to an internal used unicode representation. The enum is convertible to std::char and slint::SharedString. Use this with slint::platform::WindowEvent to supply key events to Slintโs platform abstraction.

Example

Send an tab key press event to a window

use slint::platform::{WindowEvent, Key};
fn send_tab_pressed(window: &slint::Window) {
    window.dispatch_event(WindowEvent::KeyPressed { text: Key::Tab.into() });
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Backspace

Tab

Return

Escape

Backtab

Delete

Shift

Control

Alt

AltGr

CapsLock

ShiftR

ControlR

Meta

MetaR

Space

UpArrow

DownArrow

LeftArrow

RightArrow

F1

F2

F3

F4

F5

F6

F7

F8

F9

F10

F11

F12

F13

F14

F15

F16

F17

F18

F19

F20

F21

F22

F23

F24

Insert

Home

End

PageUp

PageDown

ScrollLock

Pause

SysReq

Stop

Menu

Back