Enum slint::platform::PlatformError

pub enum PlatformError {
    NoPlatform,
    NoEventLoopProvider,
    SetPlatformError(SetPlatformError),
    Other(String),
}

The platform encountered a fatal error.

This error typically indicates an issue with initialization or connecting to the windowing system.

This can be constructed from a String:

use slint::platform::PlatformError;
PlatformError::from(format!("Could not load resource {}", 1234));

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NoPlatform

No default platform was selected, or no platform could be initialized.

If you encounter this error, make sure to either selected trough the backend-* cargo features flags, or call platform::set_platform() before running the event loop

NoEventLoopProvider

The Slint Platform does not provide an event loop.

The Platform::run_event_loop is not implemented for the current platform.

SetPlatformError(SetPlatformError)

There is already a platform set from another thread.

Other(String)

Another platform-specific error occurred