Enum slint::platform::SetPlatformError

#[non_exhaustive]
pub enum SetPlatformError {
    AlreadySet,
}

This enum describes the different error scenarios that may occur when set_platform fails.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AlreadySet

The platform has been initialized in an earlier call to set_platform.