slint::platform::software_renderer

Struct MinimalSoftwareWindow

pub struct MinimalSoftwareWindow { /* private fields */ }

This is a minimal adapter for a Window that doesn’t have any other feature than rendering using the software renderer.

Implementations

impl MinimalSoftwareWindow

pub fn new(repaint_buffer_type: RepaintBufferType) -> Rc<MinimalSoftwareWindow>

Instantiate a new MinimalWindowAdaptor

The repaint_buffer_type parameter specify what kind of buffer are passed to the SoftwareRenderer

pub fn draw_if_needed( &self, render_callback: impl FnOnce(&SoftwareRenderer), ) -> bool

If the window needs to be redrawn, the callback will be called with the renderer that should be used to do the drawing.

SoftwareRenderer::render() or SoftwareRenderer::render_by_line() should be called in that callback.

Return true if something was redrawn.