slint

Struct BackendSelector

pub struct BackendSelector { /* private fields */ }

Use the BackendSelector to configure one of Slint’s built-in backends with a renderer to accommodate specific needs of your application. This is a programmatic substitute for the SLINT_BACKEND environment variable.

For example, to configure Slint to use a renderer that supports OpenGL ES 3.0, configure the BackendSelector as follows:

let selector = BackendSelector::new().require_opengl_es_with_version(3, 0);
if let Err(err) = selector.select() {
    eprintln!("Error selecting backend with OpenGL ES support: {err}");
}

Implementations

impl BackendSelector

pub fn new() -> BackendSelector

Creates a new BackendSelector.

pub fn require_opengl_es_with_version( self, major: u8, minor: u8, ) -> BackendSelector

Adds the requirement to the selector that the backend must render with OpenGL ES and the specified major and minor version.

pub fn require_opengl_es(self) -> BackendSelector

Adds the requirement to the selector that the backend must render with OpenGL ES.

pub fn require_opengl(self) -> BackendSelector

Adds the requirement to the selector that the backend must render with OpenGL.

pub fn require_opengl_with_version( self, major: u8, minor: u8, ) -> BackendSelector

Adds the requirement to the selector that the backend must render with OpenGL and the specified major and minor version.

pub fn require_metal(self) -> BackendSelector

Adds the requirement to the selector that the backend must render with Apple’s Metal framework.

pub fn require_vulkan(self) -> BackendSelector

Adds the requirement to the selector that the backend must render with Vulkan.

pub fn require_d3d(self) -> BackendSelector

Adds the requirement to the selector that the backend must render with Direct 3D.

pub fn require_wgpu_26( self, configuration: WGPUConfiguration, ) -> BackendSelector

Adds the requirement to the selector that the backend must render using WGPU. Use this when you integrate other WGPU-based renderers with a Slint UI.

Note: This function is behind the unstable-wgpu-26 feature flag and may be removed or changed in future minor releases, as new major WGPU releases become available.

See also the slint::wgpu_26 module.

pub fn require_wgpu_27( self, configuration: WGPUConfiguration, ) -> BackendSelector

Adds the requirement to the selector that the backend must render using WGPU. Use this when you integrate other WGPU-based renderers with a Slint UI.

Note: This function is behind the unstable-wgpu-27 feature flag and may be removed or changed in future minor releases, as new major WGPU releases become available.

See also the slint::wgpu_27 module.

pub fn with_winit_window_attributes_hook( self, hook: impl Fn(WindowAttributes) -> WindowAttributes + 'static, ) -> BackendSelector

Configures this builder to use the specified winit hook that will be called before a Window is created.

It can be used to adjust settings of window that will be created.

Example

let mut backend = slint::BackendSelector::new()
    .with_winit_window_attributes_hook(|attributes| attributes.with_content_protected(true))
    .select()
    .unwrap();

Note: This function is behind the unstable-winit-030 feature flag and may be removed or changed in future minor releases, as new major Winit releases become available.

See also the slint::winit_030 module

pub fn with_winit_event_loop_builder( self, event_loop_builder: EventLoopBuilder<SlintEvent>, ) -> BackendSelector

Configures this builder to use the specified winit event loop builder when creating the event loop.

Note: This function is behind the unstable-winit-030 feature flag and may be removed or changed in future minor releases, as new major Winit releases become available.

See also the slint::winit_030 module

pub fn with_winit_custom_application_handler( self, custom_application_handler: impl CustomApplicationHandler + 'static, ) -> BackendSelector

Configures this builder to invoke the functions on the supplied application handler whenever winit wakes up the event loop.

Note: This function is behind the unstable-winit-030 feature flag and may be removed or changed in future minor releases, as new major Winit releases become available.

See also the slint::winit_030 module

pub fn with_libinput_event_hook( self, event_hook: impl Fn(&Event) -> bool + 'static, ) -> BackendSelector

Configures this builder to use the specified libinput event filter hook when the LinuxKMS backend is selected.

The provided hook is invoked for every event received. If the function returns true, the event is not dispatched further.

Note: This function is behind the unstable-libinput-09 feature flag and may be removed or changed in future minor releases, as new major Winit releases become available.

pub fn renderer_name(self, name: String) -> BackendSelector

Adds the requirement that the selected renderer must match the given name. This is equivalent to setting the SLINT_BACKEND=name environment variable and requires that the corresponding renderer feature is enabled. For example, to select the Skia renderer, enable the renderer-skia feature and call this function with skia as argument.

pub fn backend_name(self, name: String) -> BackendSelector

Adds the requirement that the selected backend must match the given name. This is equivalent to setting the SLINT_BACKEND=name environment variable and requires that the corresponding backend feature is enabled. For example, to select the winit backend, enable the backend-winit feature and call this function with winit as argument.

pub fn select(self) -> Result<(), PlatformError>

Completes the backend selection process and tries to combine with specified requirements with the different backends and renderers enabled at compile time. On success, the selected backend is automatically set to be active. Returns an error if the requirements could not be met.