Struct slint::ModelNotify

pub struct ModelNotify { /* private fields */ }

Dispatch notifications from a Model to one or several ModelPeer. Typically, you would want to put this in the implementation of the Model

Implementations

impl ModelNotify

pub fn row_changed(&self, row: usize)

Notify the peers that a specific row was changed

pub fn row_added(&self, index: usize, count: usize)

Notify the peers that rows were added

pub fn row_removed(&self, index: usize, count: usize)

Notify the peers that rows were removed

pub fn reset(&self)

Notify the peer that the model has been changed in some way and everything needs to be reloaded