Struct slint::SharedPixelBuffer

#[repr(C)]
pub struct SharedPixelBuffer<Pixel> { /* private fields */ }

SharedPixelBuffer is a container for storing image data as pixels. It is internally reference counted and cheap to clone.

You can construct a new empty shared pixel buffer with SharedPixelBuffer::new, or you can clone it from an existing contiguous buffer that you might already have, using SharedPixelBuffer::clone_from_slice.

See the documentation for Image for examples how to use this type to integrate Slint with external rendering functions.

Implementations

impl<Pixel> SharedPixelBuffer<Pixel>

pub fn width(&self) -> u32

Returns the width of the image in pixels.

pub fn height(&self) -> u32

Returns the height of the image in pixels.

pub fn size(&self) -> Size2D<u32, UnknownUnit>

Returns the size of the image in pixels.

pub fn stride(&self) -> u32

Returns the number of pixels per line.

impl<Pixel> SharedPixelBuffer<Pixel> where
Pixel: Clone,

pub fn make_mut_slice(&mut self) -> &mut [Pixel]

Return a mutable slice to the pixel data. If the SharedPixelBuffer was shared, this will make a copy of the buffer.

impl<Pixel> SharedPixelBuffer<Pixel> where
Pixel: Clone + Pod,
[Pixel]: ComponentBytes<u8>,

pub fn as_bytes(&self) -> &[u8]

Returns the pixels interpreted as raw bytes.

pub fn make_mut_bytes(&mut self) -> &mut [u8]

Returns the pixels interpreted as raw bytes.

impl<Pixel> SharedPixelBuffer<Pixel>

pub fn as_slice(&self) -> &[Pixel]

Return a slice to the pixel data.

impl<Pixel> SharedPixelBuffer<Pixel> where
Pixel: Clone + Default,

pub fn new(width: u32, height: u32) -> SharedPixelBuffer<Pixel>

Creates a new SharedPixelBuffer with the given width and height. Each pixel will be initialized with the value that Default::default() returns for the Pixel type.

impl<Pixel> SharedPixelBuffer<Pixel> where
Pixel: Clone,

pub fn clone_from_slice<SourcePixelType>(
pixel_slice: &[SourcePixelType],
width: u32,
height: u32
) -> SharedPixelBuffer<Pixel> where
[SourcePixelType]: AsPixels<Pixel>,

Creates a new SharedPixelBuffer by cloning and converting pixels from an existing slice. This function is useful when another crate was used to allocate an image and you would like to convert it for use in Slint.