Struct slint::VecModel

pub struct VecModel<T> { /* private fields */ }

A Model backed by a Vec<T>

Implementations

impl<T> VecModel<T>
where T: 'static,

pub fn from_slice(slice: &[T]) -> ModelRc<T>
where T: Clone,

Allocate a new model from a slice

pub fn push(&self, value: T)

Add a row at the end of the model

pub fn insert(&self, index: usize, value: T)

Inserts a row at position index. All rows after that are shifted. This function panics if index is > row_count().

pub fn remove(&self, index: usize) -> T

Remove the row at the given index from the model

Returns the removed row

pub fn set_vec(&self, new: impl Into<Vec<T>>)

Replace inner Vec with new data

pub fn extend<I>(&self, iter: I)
where I: IntoIterator<Item = T>,

Extend the model with the content of the iterator

Similar to Vec::extend

impl<T> VecModel<T>
where T: Clone + 'static,

pub fn extend_from_slice(&self, src: &[T])

Appends all the elements in the slice to the model

Similar to Vec::extend_from_slice