Struct slint::Window

#[repr(transparent)]
pub struct Window(_);

This type represents a window towards the windowing system, that’s used to render the scene of a component. It provides API to control windowing system specific aspects such as the position on the screen.

Implementations

impl Window

pub fn show(&self)

Registers the window with the windowing system in order to make it visible on the screen.

pub fn hide(&self)

De-registers the window from the windowing system, therefore hiding it.

pub fn set_rendering_notifier(
&self,
callback: impl FnMut(RenderingState, &GraphicsAPI<'_>) + 'static
) -> Result<(), SetRenderingNotifierError>

This function allows registering a callback that’s invoked during the different phases of rendering. This allows custom rendering on top or below of the scene.

pub fn on_close_requested(
&self,
callback: impl FnMut() -> CloseRequestResponse + 'static
)

This function allows registering a callback that’s invoked when the user tries to close a window. The callback has to return a CloseRequestResponse.

pub fn request_redraw(&self)

This function issues a request to the windowing system to redraw the contents of the window.

pub fn scale_factor(&self) -> f32

This function returns the scale factor that allows converting between logical and physical pixels.

pub fn position(&self) -> PhysicalPosition

Returns the position of the window on the screen, in physical screen coordinates and including a window frame (if present).

pub fn set_position(&self, position: impl Into<WindowPosition>)

Sets the position of the window on the screen, in physical screen coordinates and including a window frame (if present). Note that on some windowing systems, such as Wayland, this functionality is not available.

pub fn size(&self) -> PhysicalSize

Returns the size of the window on the screen, in physical screen coordinates and excluding a window frame (if present).

pub fn set_size(&self, size: impl Into<WindowSize>)

Resizes the window to the specified size on the screen, in physical pixels and excluding a window frame (if present).

pub fn dispatch_event(&self, event: WindowEvent)

Dispatch a window event to the scene.

Use this when you’re implementing your own backend and want to forward user input events.

Any position fields in the event must be in the logical pixel coordinate system relative to the top left corner of the window.

pub fn has_active_animations(&self) -> bool

Returns true if there is an animation currently active on any property in the Window; false otherwise.