slint::wgpu_24::wgpu::core::naga::back

Module glsl

Available on crate feature unstable-wgpu-24 only.

Backend for GLSL (OpenGL Shading Language).

The main structure is Writer, it maintains internal state that is used to output a Module into glsl

Supported versions

Core

330

400

410

420

430

450

ES

300

310

Structs

Features

Structure used to encode additions to GLSL that aren’t supported by all versions.

Options

Configuration used in the Writer.

PipelineOptions

A subset of options meant to be changed per pipeline.

PushConstantItem

All information to bind a single uniform value to the shader.

ReflectionInfo

Reflection info for texture mappings and uniforms.

TextureMapping

Mapping between a texture and its sampler, if it exists.

VaryingLocation

Writer

Writer responsible for all code generation.

WriterFlags

Configuration flags for the Writer.

Enums

Error

A GLSL compilation error.

Version

A GLSL version.

Constants

FIRST_INSTANCE_BINDING

SUPPORTED_CORE_VERSIONS

List of supported core GLSL versions.

SUPPORTED_ES_VERSIONS

List of supported es GLSL versions.

Type Aliases

BindingMap

Mapping between resources and bindings.