slint::wgpu_24::wgpu::core::naga

Struct Arena

pub struct Arena<T> { /* private fields */ }

Available on crate feature unstable-wgpu-24 only.

An arena holding some kind of component (e.g., type, constant, instruction, etc.) that can be referenced.

Adding new items to the arena produces a strongly-typed Handle. The arena can be indexed using the given handle to obtain a reference to the stored item.

Implementations

impl<T> Arena<T>

pub const fn new() -> Arena<T>

Create a new arena with no initial capacity allocated.

pub fn into_inner(self) -> Vec<T>

Extracts the inner vector.

pub fn len(&self) -> usize

Returns the current number of items stored in this arena.

pub fn is_empty(&self) -> bool

Returns true if the arena contains no elements.

pub fn iter(&self) -> impl DoubleEndedIterator

Returns an iterator over the items stored in this arena, returning both the itemโs handle and a reference to it.

pub fn drain(&mut self) -> impl DoubleEndedIterator

Drains the arena, returning an iterator over the items stored.

pub fn iter_mut(&mut self) -> impl DoubleEndedIterator

Returns a iterator over the items stored in this arena, returning both the itemโs handle and a mutable reference to it.

pub fn append(&mut self, value: T, span: Span) -> Handle<T>

Adds a new value to the arena, returning a typed handle.

pub fn fetch_if<F>(&self, fun: F) -> Option<Handle<T>>
where F: Fn(&T) -> bool,

Fetch a handle to an existing type.

pub fn fetch_if_or_append<F>( &mut self, value: T, span: Span, fun: F, ) -> Handle<T>
where F: Fn(&T, &T) -> bool,

Adds a value with a custom check for uniqueness: returns a handle pointing to an existing element if the check succeeds, or adds a new element otherwise.

pub fn fetch_or_append(&mut self, value: T, span: Span) -> Handle<T>
where T: PartialEq,

Adds a value with a check for uniqueness, where the check is plain comparison.

pub fn try_get(&self, handle: Handle<T>) -> Result<&T, BadHandle>

pub fn get_mut(&mut self, handle: Handle<T>) -> &mut T

Get a mutable reference to an element in the arena.

pub fn range_from(&self, old_length: usize) -> Range<T>

Get the range of handles from a particular number of elements to the end.

pub fn clear(&mut self)

Clears the arena keeping all allocations

pub fn get_span(&self, handle: Handle<T>) -> Span

pub fn check_contains_handle(&self, handle: Handle<T>) -> Result<(), BadHandle>

Assert that handle is valid for this arena.

pub fn check_contains_range( &self, range: &Range<T>, ) -> Result<(), BadRangeError>

Assert that range is valid for this arena.