slint::wgpu_26::wgpu

Type Alias ShaderModuleDescriptorHlsl

pub type ShaderModuleDescriptorHlsl<'a> = ShaderModuleDescriptorHlsl<'a, Option<&'a str>>;

Available on crate feature unstable-wgpu-26 only.

Descriptor for a shader module given by DirectX HLSL source.

This type is unique to the Rust API of wgpu. In the WebGPU specification, only WGSL source code strings are accepted.

Aliased Type

pub struct ShaderModuleDescriptorHlsl<'a> {
    pub entry_point: String,
    pub label: Option<&'a str>,
    pub num_workgroups: (u32, u32, u32),
    pub source: &'a str,
}

Fields
entry_point: String

Entrypoint.

label: Option<&'a str>

Debug label of the shader module. This will show up in graphics debuggers for easy identification.

num_workgroups: (u32, u32, u32)

Number of workgroups in each dimension x, y and z.

source: &'a str

Shader HLSL source.