slint::wgpu_26::wgpu::wgc

Module binding_model

Available on crate feature unstable-wgpu-26 only.

Structs

BindGroup

BindGroupDescriptor

Describes a group of bindings and the resources to be bound.

BindGroupDynamicBindingData

BindGroupEntry

Bindable resource and the slot to bind it to. cbindgen:ignore

BindGroupLayout

Bind group layout.

BindGroupLayoutDescriptor

Describes a BindGroupLayout.

BindingTypeMaxCountError

BufferBinding

LateMinBufferBindingSizeMismatch

PipelineLayout

PipelineLayoutDescriptor

Describes a pipeline layout.

Enums

BindError

BindGroupLayoutEntryError

BindingResource

BindingTypeMaxCountErrorKind

BindingZone

CreateBindGroupError

CreateBindGroupLayoutError

CreatePipelineLayoutError

GetBindGroupLayoutError

PushConstantUploadError

Type Aliases

ResolvedBindGroupDescriptor

cbindgen:ignore

ResolvedBindGroupEntry

cbindgen:ignore

ResolvedBindingResource

ResolvedBufferBinding

ResolvedPipelineLayoutDescriptor

cbindgen:ignore