slint::wgpu_26::wgpu::wgc

Module resource

Available on crate feature unstable-wgpu-26 only.

Structs

Blas

Buffer

BufferMapOperation

DestroyedBuffer

A buffer that has been marked as destroyed and is staged for actual deletion soon.

DestroyedResourceError

DestroyedTexture

A texture that has been marked as destroyed and is staged for actual deletion soon.

FlushedStagingBuffer

InvalidResourceError

MissingBufferUsageError

MissingTextureUsageError

QuerySet

ResourceErrorIdent

Sampler

SamplerDescriptor

Describes a Sampler

StagingBuffer

A temporary buffer, consumed by the command that uses it.

Texture

TextureView

TextureViewDescriptor

Describes a TextureView.

Tlas

Enums

BufferAccessError

CreateBufferError

CreateQuerySetError

CreateSamplerError

CreateTextureError

CreateTextureViewError

Fallible

SamplerFilterErrorType

TextureClearMode

TextureDimensionError

TextureErrorDimension

TextureViewDestroyError

TextureViewNotRenderableReason

Traits

Labeled

ParentDevice

RawResourceAccess

Allow access to the hal resource as guarded by the SnatchGuard.

ResourceType

Type Aliases

BlasCompactCallback

BlasDescriptor

BlasPrepareCompactResult

BufferAccessResult

BufferDescriptor

BufferMapCallback

QuerySetDescriptor

TextureDescriptor

TlasDescriptor