slint::wgpu_26::wgpu::wgc::resource

Trait Labeled

pub trait Labeled: ResourceType {
    // Required method
    fn label(&self) -> &str;

    // Provided method
    fn error_ident(&self) -> ResourceErrorIdent { ... }
}

Available on crate feature unstable-wgpu-26 only.

Required Methods

fn label(&self) -> &str

Returns a string identifying this resource for logging and errors.

It may be a user-provided string or it may be a placeholder from wgpu.

It is non-empty unless the user-provided string was empty.

Provided Methods

fn error_ident(&self) -> ResourceErrorIdent

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.