slint::wgpu_26::wgpu::wgc

Module validation

Available on crate feature unstable-wgpu-26 only.

Structs

Interface

InterfaceVar

NumericType

Enums

BindingError

BindingLayoutSource

BindingTypeName

FilteringError

InputError

StageError

Errors produced when validating a programmable stage of a pipeline.

Functions

check_texture_format

Return true if the fragment format is covered by the provided output.

map_storage_format_from_naga

map_storage_format_to_naga

validate_color_attachment_bytes_per_sample

Type Aliases

StageIo