slint::wgpu_27::wgpu::wgc

Module command

Available on crate feature unstable-wgpu-27 only.

Command Encoding

TODO: High-level description of command encoding.

The convention in this module is that functions accepting a [&mut dyn hal::DynCommandEncoder] are low-level helpers and may assume the encoder is in the open state, ready to encode commands. Encoders that are not open should be nested within some other container that provides additional state tracking, like [InnerCommandEncoder].

Modules

bundle_ffi

ffi

Types that are useful for FFI bindings to wgpu.

Structs

ArcRenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a render pass.

CommandBuffer

A buffer of commands to be submitted to the GPU for execution.

CommandBufferMutable

The mutable state of a CommandBuffer.

ComputePass

A pass’s encoder state and its validity are two distinct conditions, i.e., the full matrix of (open, ended) x (valid, invalid) is possible.

ComputePassDescriptor

ComputePassError

Error encountered when performing a compute pass, stored for later reporting when encoding ends.

PassChannel

Describes an individual channel within a render pass, such as color, depth, or stencil.

PassStateError

Variant of EncoderStateError that includes the pass scope.

PassTimestampWrites

Describes the writing of timestamp values in a render or compute pass.

Rect

RenderBundle

RenderBundleEncoder

RenderBundleEncoderDescriptor

Describes a RenderBundleEncoder.

RenderBundleError

Error encountered when finishing recording a render bundle.

RenderPass

A pass’s encoder state and its validity are two distinct conditions, i.e., the full matrix of (open, ended) x (valid, invalid) is possible.

RenderPassColorAttachment

Describes a color attachment to a render pass.

RenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a render pass.

RenderPassDescriptor

Describes the attachments of a render pass.

RenderPassError

Error encountered when performing a render pass.

Enums

ArcCommand

ArcComputeCommand

Equivalent to ComputeCommand but the Ids resolved into resource Arcs.

AttachmentError

AttachmentErrorLocation

Describes an attachment location in words.

ClearError

Error encountered while attempting a clear.

ColorAttachmentError

Command

CommandEncoderError

ComputeCommand

ComputePassErrorInner

Error encountered when performing a compute pass.

CopySide

CreateRenderBundleError

Error type returned from RenderBundleEncoder::new if the sample count is invalid.

DebugGroupError

DispatchError

DrawCommandFamily

The type of draw command(indexed or not, or mesh shader)

DrawError

Error validating a draw call.

DrawKind

EncoderStateError

Errors related to the state of a command or pass encoder.

ExecutionError

Error type returned from RenderBundleEncoder::new if the sample count is invalid.

LoadOp

Operation to perform to the output attachment at the start of a render pass.

PassErrorScope

A command that can be recorded in a pass or bundle.

QueryError

Error encountered when dealing with queries

QueryUseError

Error encountered while trying to use queries

RenderCommandError

Error encountered when encoding a render command. This is the shared error set between render bundles and passes.

RenderPassErrorInner

Error encountered when performing a render pass.

ResolveError

Error encountered while trying to resolve a query.

ResolvedPassChannel

SimplifiedQueryType

StoreOp

Operation to perform to the output attachment at the end of a render pass.

TimestampWritesError

TransferError

Error encountered while attempting a data transfer.

Type Aliases

ArcRenderPassColorAttachment

ArcRenderPassColorAttachmentArray

ComputeBasePass

CopyExternalImageDestInfo

cbindgen:ignore

RenderBasePass

RenderBundleDescriptor

TexelCopyBufferInfo

cbindgen:ignore

TexelCopyTextureInfo

cbindgen:ignore