slint::wgpu_27::wgpu::wgt::instance

Struct NoopBackendOptions

pub struct NoopBackendOptions {
    pub enable: bool,
}

Available on crate feature unstable-wgpu-27 only.

Configuration for the noop backend.

Part of BackendOptions.

Fields
enable: bool

Whether to allow the noop backend to be used.

The noop backend stubs out all operations except for buffer creation and mapping, so it must not be used when not expected. Therefore, it will not be used unless explicitly enabled.

Implementations

impl NoopBackendOptions

pub fn from_env_or_default() -> NoopBackendOptions

Choose whether the noop backend is enabled from the environment.

It will be enabled if the environment variable WGPU_NOOP_BACKEND has the value 1 and not otherwise. Future versions may assign other meanings to other values.

pub fn with_env(self) -> NoopBackendOptions

Takes the given options, modifies them based on the environment variables, and returns the result.

See from_env_or_default() for the interpretation.