Module slint::docs

This is a pseudo module which only exist for documentation purposes as a way to show the Slint documentation as part of rustdoc.

The generated_code module contains an commented example of what is generated from the .slint file

The langref module is the reference documentation for the .slint language.

The widgets and builtin_elements modules contains the documentation of elements usable within the .slint files

The layouting module contains the documentation to position the elements.

Modules

builtin_elements

Builtin Elements

builtin_enums

debugging_techniques

Debugging Techniques

generated_code

This module exists only to explain the API of the code generated from .slint design markup. Its described structure is not really contained in the compiled crate.

langref

The .slint language reference

layouting

Positioning and Layout of Elements

mcu

Slint on Microcontrollers

recipes

Recipes and Examples

widgets

Widgets