Module slint::platform

This module contains items that you need to use or implement if you want use Slint in an environment without one of the supplied platform backends such as qt or winit.

The primary interface is the platform::Platform trait. Pass your implementation of it to Slint by calling platform::set_platform() early on in your application, before creating any Slint components.

The Slint on Microcontrollers documentation has additional examples.

Modules

software_renderer

This module contains the SoftwareRenderer and related types.

Enums

SetPlatformError

This enum describes the different error scenarios that may occur when set_platform fails.

Traits

EventLoopProxy

Trait that is returned by the Platform::new_event_loop_proxy

Platform

This trait defines the interface between Slint and platform APIs typically provided by operating and windowing systems.

WindowAdapter

This trait represents the adaptation layer between the Window API, and the internal type from the backend that provides functionality such as device-independent pixels, window resizing, and other typically windowing system related tasks.

Functions

duration_until_next_timer_update

Returns the duration before the next timer is expected to be activated. This is the largest amount of time that you can wait before calling update_timers_and_animations().

set_platform

Set the Slint platform abstraction.

update_timers_and_animations

Call this function to update and potentially activate any pending timers, as well as advance the state of any active animtaions.