Module slint::platform::software_renderer

This module contains the SoftwareRenderer and related types.

It is only enabled when the renderer-software Slint feature is enabled.

Structs

MinimalSoftwareWindow

This is a minimal adapter for a Window that doesn’t have any other feature than rendering using the software renderer.

PhysicalRegion

Represents a rectangular region on the screen, used for partial rendering.

PremultipliedRgbaColor

A color whose component have been pre-multiplied by alpha

Rgb565Pixel

A 16bit pixel that has 5 red bits, 6 green bits and 5 blue bits

SoftwareRenderer

A Renderer that do the rendering in software

Enums

RepaintBufferType

This enum describes which parts of the buffer passed to the SoftwareRenderer may be re-used to speed up painting.

Traits

LineBufferProvider

This trait defines a bi-directional interface between Slint and your code to send lines to your screen, when using the SoftwareRenderer::render_by_line function.

TargetPixel

Trait for the pixels in the buffer