slint::platform::software_renderer

Trait TargetPixel

pub trait TargetPixel: Sized + Copy {
    // Required methods
    fn blend(&mut self, color: PremultipliedRgbaColor);
    fn from_rgb(red: u8, green: u8, blue: u8) -> Self;

    // Provided methods
    fn blend_slice(slice: &mut [Self], color: PremultipliedRgbaColor) { ... }
    fn background() -> Self { ... }
}

Trait for the pixels in the buffer

Required Methods

fn blend(&mut self, color: PremultipliedRgbaColor)

Blend a single pixel with a color

fn from_rgb(red: u8, green: u8, blue: u8) -> Self

Create a pixel from the red, gree, blue component in the range 0..=255

Provided Methods

fn blend_slice(slice: &mut [Self], color: PremultipliedRgbaColor)

Blend a color to all the pixel in the slice.

fn background() -> Self

Pixel which will be filled as the background in case the slint view has transparency

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.