Struct slint::Color

#[repr(C)]
pub struct Color { /* private fields */ }

Color represents a color in the Slint run-time, represented using 8-bit channels for red, green, blue and the alpha (opacity). It can be conveniently converted using the to_ and from_ (a)rgb helper functions:

let col = some_color.to_argb_f32();
do_something_with_red_and_green(col.red, col.green);

let RgbaColor { red, blue, green, .. } = some_color.to_argb_u8();
do_something_with_red(red);

let new_col = Color::from(RgbaColor{ red: 0.5, green: 0.65, blue: 0.32, alpha: 1.});

Implementations

impl Color

pub const fn from_argb_encoded(encoded: u32) -> Color

Construct a color from an integer encoded as 0xAARRGGBB

pub fn as_argb_encoded(&self) -> u32

Returns (alpha, red, green, blue) encoded as u32

pub const fn from_argb_u8(alpha: u8, red: u8, green: u8, blue: u8) -> Color

Construct a color from the alpha, red, green and blue color channel parameters.

pub const fn from_rgb_u8(red: u8, green: u8, blue: u8) -> Color

Construct a color from the red, green and blue color channel parameters. The alpha channel will have the value 255.

pub fn from_argb_f32(alpha: f32, red: f32, green: f32, blue: f32) -> Color

Construct a color from the alpha, red, green and blue color channel parameters.

pub fn from_rgb_f32(red: f32, green: f32, blue: f32) -> Color

Construct a color from the red, green and blue color channel parameters. The alpha channel will have the value 255.

pub fn to_argb_u8(&self) -> RgbaColor<u8>

Converts this color to an RgbaColor struct for easy destructuring.

pub fn to_argb_f32(&self) -> RgbaColor<f32>

Converts this color to an RgbaColor struct for easy destructuring.

pub fn red(self) -> u8

Returns the red channel of the color as u8 in the range 0..255.

pub fn green(self) -> u8

Returns the green channel of the color as u8 in the range 0..255.

pub fn blue(self) -> u8

Returns the blue channel of the color as u8 in the range 0..255.

pub fn alpha(self) -> u8

Returns the alpha channel of the color as u8 in the range 0..255.

pub fn brighter(&self, factor: f32) -> Color

Returns a new version of this color that has the brightness increased by the specified factor. This is done by converting the color to the HSV color space and multiplying the brightness (value) with (1 + factor). The result is converted back to RGB and the alpha channel is unchanged. So for example brighter(0.2) will increase the brightness by 20%, and calling brighter(-0.5) will return a color that’s 50% darker.

pub fn darker(&self, factor: f32) -> Color

Returns a new version of this color that has the brightness decreased by the specified factor. This is done by converting the color to the HSV color space and dividing the brightness (value) by (1 + factor). The result is converted back to RGB and the alpha channel is unchanged. So for example darker(0.3) will decrease the brightness by 30%.