slint::wgpu_27::wgpu::wgc::command

Type Alias ComputeBasePass

pub type ComputeBasePass = BasePass<ArcComputeCommand, ComputePassError>;

Available on crate feature unstable-wgpu-27 only.

Aliased Type

pub struct ComputeBasePass {
    pub label: Option<String>,
    pub error: Option<ComputePassError>,
    pub commands: Vec<ArcComputeCommand>,
    pub dynamic_offsets: Vec<u32>,
    pub string_data: Vec<u8>,
    pub push_constant_data: Vec<u32>,
}

Fields
label: Option<String>error: Option<ComputePassError>

If the pass is invalid, contains the error that caused the invalidation.

If the pass is valid, this is None.

Passes are serialized into traces. but we don’t support doing so for passes containing errors. These serde attributes allow E to be Infallible.

commands: Vec<ArcComputeCommand>

The stream of commands.

The commands are moved out of this vector when the pass is ended (i.e. at the same time that parent is taken out of the ComputePass/RenderPass).

dynamic_offsets: Vec<u32>

Dynamic offsets consumed by [SetBindGroup] commands in commands.

Each successive SetBindGroup consumes the next [num_dynamic_offsets] values from this list.

string_data: Vec<u8>

Strings used by debug instructions.

Each successive [PushDebugGroup] or [InsertDebugMarker] instruction consumes the next len bytes from this vector.

push_constant_data: Vec<u32>

Data used by SetPushConstant instructions.

See the documentation for [RenderCommand::SetPushConstant] and [ComputeCommand::SetPushConstant] for details.